backend/requirements.txt
fastapi
mangum
pymongo
python-jose[cryptography]
bcrypt
python-dotenv
uvicorn[standard]
pydantic
python-multipart